% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{avg_pwr}
\alias{avg_pwr}
\title{Calculates the average power of a FDR controlling procedure}
\usage{
avg_pwr(b, Fn_alt, m0, m, parallel = TRUE, progress = TRUE,
  quick = FALSE)
}
\arguments{
\item{b}{Vector of upper bounds b}

\item{Fn_alt}{Distribution (cdf) under the alternative}

\item{m0}{The number of true hypotheses - Can be at most m}

\item{m}{The number of hypotheses - Can be at most length(b)}

\item{parallel}{Use parallelization to speed up the calculation?}

\item{progress}{Show a progress bar?}

\item{quick}{Use lower precision to speed up calculation?}
}
\value{
Returns the average power
}
\description{
Calculates, under the assumption of joint stochastic independence, the average power
E[(R-V)/(m-m_0)] of a FDR controlling step-up test given by the vector of thresholds b.
}
\details{
Calculates, under the model FM(m,m0,Fn_alt) in https://arxiv.org/abs/1812.09063 the average power E[(R-V)/(m-m_0)].
For the definition of R and V see the documentation of \code{dfd_fm}.

The implementation is based on a generalization of Noe's recursion. For the details see https://arxiv.org/abs/1812.09063.
}
\examples{
#Equation (24) in Glueck, D., Mandel, J., Karimpour-Fard, A., et al. (2008). Exact Calculations of Average Power for the Benjamini-Hochberg Procedure. The International Journal of Biostatistics, 4(1), doi:10.2202/1557-4679.1103
Fn <- function(t) 1 + pnorm(qnorm(t/2)-sqrt(N)) - pnorm(qnorm(1-t/2)-sqrt(N))
#Setting of Table 2 of Glueck (2008)
N <- 5
m <- 5
m0 <- 4
t <- 0.05 * (1:m)/m
#Calculate the last power value in Table 2 of Glueck (2008)
print(OrdStat::avg_pwr(t,Fn,m0,m,progress=F))

}
