% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_local_level_gof_test.R
\name{calibrate_local_level_gof_test}
\alias{calibrate_local_level_gof_test}
\title{Calibrates the the goodness-of-fit test with equal local levels proposed by Gontscharuk et al. (2016)}
\usage{
calibrate_local_level_gof_test(n, alpha)
}
\arguments{
\item{n}{The number of hypotheses}

\item{alpha}{The desired global level}
}
\value{
Returns the local level necessary to achieve the desired global level
}
\description{
Calibrates the the goodness-of-fit test with equal local levels proposed by
 Gontscharuk, V., Landwehr, S., Finner, H., (2016). Goodness of fit tests in
 terms of local levels with special emphasis on higher criticism tests.
 Bernoulli 22, 1331-1363. doi:10.3150/14-BEJ694.
}
