% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambda_pwr}
\alias{lambda_pwr}
\title{Calculates the lambda-power of a FDR controlling procedure}
\usage{
lambda_pwr(b, Fn_alt, pr, m, lambda, parallel = TRUE, progress = TRUE,
  quick = FALSE)
}
\arguments{
\item{b}{Vector of upper bounds b}

\item{Fn_alt}{Distribution (cdf) under the alternative}

\item{pr}{The probability that a given hypothesis is true}

\item{m}{The number of hypotheses - Can be at most length(b)}

\item{lambda}{The threshold lambda}

\item{parallel}{Use parallelization to speed up the calculation?}

\item{progress}{Show a progress bar?}

\item{quick}{Use lower precision to speed up calculation?}
}
\value{
Returns the lambda power
}
\description{
Calculates, under the model RM(m,pr,Fn_alt) in https://arxiv.org/abs/1812.09063 the lambda-power P[(R-V)/(m-M_0)>=lambda].
}
\details{
Calculates, under the assumption of joint stochastic independence, the lambda-power
P[(R-V)/(m-M_0)>=lambda] of a FDR controlling step-up test given by the vector of thresholds b
and where M_0 is binomially distributed pbinom(.,m,pr).
For the definition of R and V see the documentation of \code{dfd_fm}.

The implementation is based on a generalization of Noe's recursion. For the details see https://arxiv.org/abs/1812.09063.
}
\examples{
#Calculate the lambda power for the setting given by the first column in Table 3
# of Izmirlian, G., (2018). Average power and lambda-power in multiple testing
# scenarios when the benjamini-hochberg false discovery rate procedure is used. arXiv preprint arXiv:1801.03989
lambda <- 0.9
m <- 200
n <- 70
r <- 5/200
alpha <- 0.15
theta <- 0.6
df <- 2*n-2
ncp <- sqrt(n/2) * theta
t <- alpha * (1:m)/m
pr <- 1-r
Fn <- function(t) pt(qt(t/2,df,0,F),df,ncp,F)-pt(-qt(t/2,df,0,F),df,ncp)
print(OrdStat::lambda_pwr(t,Fn,pr,m,lambda))

}
