% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{noe_faithful}
\alias{noe_faithful}
\title{Joint distribution of order statistics for the one- or two-group case}
\usage{
noe_faithful(v1, v2, n1 = -1L, n2 = -1L, parallel = TRUE)
}
\arguments{
\item{v1}{Boundaries for the first group G_1(b)}

\item{v2}{Boundaries for the second group G_2(b)}

\item{n1}{The maximal size of the first group - Can be at most length(v1)}

\item{n2}{The maximal size of the second group - Can be at most length(v2)}

\item{parallel}{Use parallelization to speed up the calculation?}
}
\description{
Calculates, under the assumption of joint stochastic independence, the joint distribution of order statistics for the one- or two-group case.
}
\details{
Calculates, under the assumption of joint stochastic independence, the probability P(X_(1:n)<= b_1,...,X_(n:n)<= b_n)
where X_(i:n) denotes the i-th order statistic of the sample X_1,...,X_n and the X_k belong to two groups distributed according
to a cdf G_1 or G_2.

The implementation is based on a generalization of Noe's recursion.

Returns a matrix M, where M[i,j] is the above probability (for n=i-1+j-1) where (i-1) of the X_k belong to the first group
and (j-1) belong to the second group.
}
