% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pordstat2}
\alias{pordstat2}
\title{Joint distribution of order statistics for the two-group case}
\usage{
pordstat2(b, Fn_alt, n1 = -1L, n2 = -1L, parallel = TRUE,
  progress = TRUE, quick = FALSE)
}
\arguments{
\item{b}{Vector of upper bounds b}

\item{Fn_alt}{Distribution (cdf) under the alternative}

\item{n1}{The maximal size of the first group - Can be at most length(b)}

\item{n2}{The maximal size of the second group - Can be at most length(b)}

\item{parallel}{Use parallelization to speed up the calculation?}

\item{progress}{Show a progress bar?}

\item{quick}{Use lower precision to speed up calculation?}
}
\value{
Returns a matrix M, where M[i,j] is the above probability (for n=i-1+j-1) where (i-1) of the X_k belong to the first group
and (j-1) belong to the second group.
}
\description{
Calculates, under the assumption of joint stochastic independence, the joint distribution of order statistics for the one- or two-group case.
}
\details{
Calculates, under the assumption of joint stochastic independence, the probability P(X_(1:n)<= b_1,...,X_(n:n)<= b_n)
where X_(i:n) denotes the i-th order statistic of the sample X_1,...,X_n and the X_k belong to two groups.
The first is Uni[0,1] and the seconds is given by its cdf Fn_alt.

The implementation is based on a generalization of Noe's recursion. For the details see https://arxiv.org/abs/1812.09063.
}
