\name{OrdStat-package}
\alias{OrdStat-package}
\alias{OrdStat}
\docType{package}
\title{
  Calculate the joint distribution of order statistics
}
\description{
  Calculate the joint distribution of order statistics for the one- or two-group case and related quantities.
}
\details{
}
\author{
Jonathan von Schroeder

Maintainer: Jonathan von Schroeder <jvs@uni-bremen.de>
}
\references{
}
\keyword{ package }
\seealso{
}
